requirements-macos.txt
adjustText==0.8
anndata==0.8.0
cellrank==1.3.1
gseapy==1.0.4
jupyterlab
matplotlib==3.5.3
msmtools==1.2.5
#mpi4py == 3.0.3
numpy==1.21.6
pandas==1.2.4
plotly==4.14.1
pyEMMA==2.5.6
scanpy==1.9.3
scikit_learn==0.23.2
scipy==1.5.2
scvelo==0.2.4
seaborn==0.12.2
leidenalg
umap-learn==0.4.0